pyproject.toml
[project]
name = "va-prestashop"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "ipython>=8.18.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.11",
    "ollama>=0.6.1",
    "pydantic-settings>=2.11.0",
    "pyyaml>=6.0.3",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.50.0",
    "tenacity>=9.1.2",
]
